Crate hopr_path

This Rust crate contains all the path construction and path selection algorithms in the HOPR mixnet.

Modules

channel_graph

Defines the graph of HOPR payment channels.

errors

selectors

Implements different path selectors in the ChannelGraph.

Structs

ChainPath

Represents a NonEmptyPath that completely specifies a route using Addresses.

TransportPath

A NonEmptyPath that can be used to route packets using [OffchainPublicKey].

ValidatedPath

Represents NonEmptyPath that has been resolved and validated.

Enums

PathAddress

Represents a type that determines a hop on a Path.

Traits

NonEmptyPath

A Path that is guaranteed to have at least one hop - the destination.

Path

Base implementation of an abstract path.

PathAddressResolver

Allows resolution of [OffchainPublicKey] for a given [Address] or vice versa.

Type Aliases

ChannelPath