hopr_path::channel_graph

Struct ChannelGraph

pub struct ChannelGraph { /* private fields */ }

Implements a HOPR payment channel graph (directed) cached in-memory.

This structure is useful for tracking channel state changes and packet path finding.

The edges are updated only from the Indexer, and therefore the graph contains only the channels seen on-chain. The nodes and their qualities are updated as they are observed on the network.

Using this structure is much faster than querying the DB and therefore is preferred for per-packet path-finding computations. Per default, the graph does not track channels in the Closed state and therefore cannot detect channel re-openings.

When a node reaches zero quality and there are no edges (channels) containing this node, it is removed from the graph entirely.

Implementations

impl ChannelGraph

pub const INTERMEDIATE_HOPS: usize = 3usize

The maximum number of intermediate hops the automatic path-finding algorithm will look for.

pub fn new(me: Address, cfg: ChannelGraphConfig) -> Self

Creates a new instance with the given self Address.

pub fn count_channels(&self) -> usize

Number of channels (edges) in the graph.

pub fn count_nodes(&self) -> usize

Number of nodes in the graph.

pub fn is_own_channel(&self, channel: &ChannelEntry) -> bool

Checks if the channel is incoming to or outgoing from this node

pub fn my_address(&self) -> Address

Convenience method to get this node’s own address

pub fn get_channel( &self, source: &Address, destination: &Address, ) -> Option<&ChannelEntry>

Looks up an Open or PendingToClose channel given the source and destination. Returns None if no such edge exists in the graph.

pub fn get_node(&self, node: &Address) -> Option<&Node>

Gets the node information. Returns None if no such node exists in the graph.

pub fn open_channels_from( &self, source: Address, ) -> impl Iterator<Item = (&Node, &ChannelEdge)>

Gets all Open outgoing channels going from the given source.

pub fn has_path(&self, source: &Address, destination: &Address) -> bool

Checks whether there is any path via Open channels that connects source and destination This does not need to be necessarily a multi-hop path.

pub fn update_channel( &mut self, channel: ChannelEntry, ) -> Option<Vec<ChannelChange>>

Inserts or updates the given channel in the channel graph. Returns a set of changes if the channel was already present in the graphs or None if the channel was not previously present in the channel graph.

pub fn update_node_score( &mut self, address: &Address, score_update: NodeScoreUpdate, )

Updates the quality of a node (inserting it into the graph if it does not exist yet), based on the given NodeScoreUpdate.

pub fn update_channel_score( &mut self, source: &Address, destination: &Address, score: f64, )

Updates the score value of network connection between source and destination The given score value must always be non-negative.

pub fn get_channel_score( &self, source: &Address, destination: &Address, ) -> Option<f64>

Gets quality of the given channel. Returns None if no such channel exists, or no quality has been set for that channel.

pub fn contains_channel(&self, channel: &ChannelEntry) -> bool

Checks whether the given channel is in the graph already.

pub fn contains_node(&self, address: &Address) -> bool

Checks whether the given node is in the channel graph.

pub fn as_dot(&self, cfg: GraphExportConfig) -> String

Outputs the channel graph in the DOT (graphviz) format with the given config.